update dashboard user

+++ app_demo_model/views.py
@@ -16,7 +16,25 @@ import numpy as np
 
 @login_required
 def upload_model(request):   
-    return render(request, 'app_demo_model/upload_model.html')
+    user = request.user
+    if (user.is_superuser == True or user.is_staff == True):
+        return render(request, 'app_demo_model/upload_model.html')
+    else:
+        return render(request, 'app_general/errors_page.html')
+        
+def condition(x):
+    if x > 3.49:
+        return 'excellent'
+    elif x > 2.99:
+        return 'very good'
+    elif x > 2.49:
+        return 'good'
+    elif x > 1.99:
+        return'medium'
+    elif x > 1.49:
+        return 'poor'
+    else:
+        return'very poor'
 
 @login_required
 def upload_sci_model(request):
@@ -36,10 +54,33 @@ def upload_sci_model(request):
         else :
             messages.info(request, "ต้องการไฟล์ของข้อมูลที่เป็น excel หรือ csv")
             return render(request, 'app_demo_model/upload_model.html')
-
+        
+        #check type column
+        for i in df.columns:
+            if df.dtypes[i] != np.object_ and df.dtypes['admission_grade'] != np.float64:
+                messages.info(request, "ในคอลัมน์ของเกรดเฉลี่ยต้องการข้อมูล excellent, very good, good, medium, poor, very poor")
+                return render(request, 'app_demo_model/upload_model.html')
+        
         print('read data')
         df = df.dropna()#delete row missing value
         # print(df.head())
+        
+        if df.dtypes['admission_grade'] == np.float64:
+            df2 = pd.DataFrame()
+            df2['major'] = df['major']
+            df2['admission_grade'] = (df['admission_grade'].apply(condition))
+            df2['gpa_year_1'] = df['gpa_year_1'].apply(condition)
+            df2['thai'] = df['thai'].apply(condition)
+            df2['math'] = df['math'].apply(condition)
+            df2['sci'] = df['sci'].apply(condition)
+            df2['society'] = df['society'].apply(condition)
+            df2['hygiene'] = df['hygiene'].apply(condition)
+            df2['art'] = df['art'].apply(condition)
+            df2['career'] = df['career'].apply(condition)
+            df2['langues'] = df['langues'].apply(condition)
+            df2['status'] = df['status']
+            df = df2
+        print(df)
         
         a = df[(df['major'] == 'ICT')|(df['major'] == 'DSSI')|(df['major'] == 'polymer')]
         # print(a.head())
@@ -51,14 +92,8 @@ def upload_sci_model(request):
         
         p = df[(df['major'] == 'bio')|(df['major'] == 'chemi')|(df['major'] == 'math')|(df['major'] == 'microBio')|(df['major'] == 'physics')]
         # print(p.head())
-        # print(len(p))
+        # print(len(p))         
         
-        #check type column
-        for i in df.columns:
-            if df.dtypes[i] != np.object_:
-                messages.info(request, "ในคอลัมน์ของเกรดเฉลี่ยต้องการข้อมูล excellent, very good, good, medium, poor, very poor")
-                return render(request, 'app_demo_model/upload_model.html')
-              
         # import_data = dataset.load(df)
         import_data = dataset.load(a)
         result = applied.import_data(dataset, dry_run=True, raise_errors=True)
@@ -82,20 +117,29 @@ def upload_sci_model(request):
 
 @login_required
 def data_in_applied_sci(request):
-    applied = AppliedScience.objects.all() #for all the records
-    # print(applied)
-    total = applied.count() 
-    context={
-      'applied':applied,
-      'total': total,
-    } 
-    return render(request, 'app_demo_model/data_in_applied_model.html', context)
+    user = request.user
+    if (user.is_superuser == True or user.is_staff == True):
+        applied = AppliedScience.objects.all() #for all the records
+        # print(applied)
+        total = applied.count() 
+        context={
+        'applied':applied,
+        'total': total,
+        } 
+        return render(request, 'app_demo_model/data_in_applied_model.html', context)
+    else:
+        return render(request, 'app_general/errors_page.html')
 
 @login_required
 def delete_data_applied(request):
-    applied = AppliedScience.objects.all()
-    applied.delete()
-    return render(request, 'app_demo_model/data_in_applied_model.html')
+    user = request.user
+    if (user.is_superuser == True or user.is_staff == True):
+        applied = AppliedScience.objects.all()
+        applied.delete()
+        return render(request, 'app_demo_model/data_in_applied_model.html')
+    else:
+        return render(request, 'app_general/errors_page.html')
+    
 
 # @login_required
 # def upload_health_sci_model(request):
@@ -130,19 +174,27 @@ def delete_data_applied(request):
 
 @login_required
 def data_in_health_sci(request):
-    health = HealthScience.objects.all() #for all the records
-    total = health.count() 
-    context={
-      'health':health,
-      'total': total,
-    } 
-    return render(request, 'app_demo_model/data_in_health_model.html', context)
+    user = request.user
+    if (user.is_superuser == True or user.is_staff == True):
+        health = HealthScience.objects.all() #for all the records
+        total = health.count() 
+        context={
+        'health':health,
+        'total': total,
+        } 
+        return render(request, 'app_demo_model/data_in_health_model.html', context)
+    else:
+        return render(request, 'app_general/errors_page.html')
 
 @login_required
 def delete_data_health(request):
-    applied = HealthScience.objects.all()
-    applied.delete()
-    return render(request, 'app_demo_model/data_in_health_model.html')
+    user = request.user
+    if (user.is_superuser == True or user.is_staff == True):
+        applied = HealthScience.objects.all()
+        applied.delete()
+        return render(request, 'app_demo_model/data_in_health_model.html')
+    else:
+        return render(request, 'app_general/errors_page.html')
 
 # @login_required
 # def upload_pure_sci_model(request):
@@ -176,49 +228,63 @@ def delete_data_health(request):
 
 @login_required
 def data_in_pure_sci(request):
-    pure = PureScience.objects.all() #for all the records
-    # print(pure)
-    total = pure.count() 
-    print(total)
-    context={
-      'pure': pure,
-      'total': total,
-    } 
-    return render(request, 'app_demo_model/data_in_pure_model.html', context)
+    user = request.user
+    if (user.is_superuser == True or user.is_staff == True):
+        pure = PureScience.objects.all() #for all the records
+        # print(pure)
+        total = pure.count() 
+        print(total)
+        context={
+        'pure': pure,
+        'total': total,
+        } 
+        return render(request, 'app_demo_model/data_in_pure_model.html', context)
+    else:
+        return render(request, 'app_general/errors_page.html')
 
 @login_required
 def delete_data_pure(request):
-    pure = PureScience.objects.all()
-    pure.delete()
-    return render(request, 'app_demo_model/data_in_pure_model.html')
+    user = request.user
+    if (user.is_superuser == True or user.is_staff == True):
+        pure = PureScience.objects.all()
+        pure.delete()
+        return render(request, 'app_demo_model/data_in_pure_model.html')
+    else:
+        return render(request, 'app_general/errors_page.html')
 
 @login_required
 def show_model(request):
-    applied = AppliedScience.objects.all()
-    health = HealthScience.objects.all()
-    pure = PureScience.objects.all()
-    total_applied = applied.count()
-    total_health = health.count()
-    total_pure = pure.count()
-    total = total_applied + total_health + total_pure
-    context = {
-        'applied': applied,
-        'health': health,
-        'pure': pure,
-        'total': total,
-    }
-    
-    return render(request, 'app_demo_model/show_model.html', context)
+    user = request.user
+    if (user.is_superuser == True or user.is_staff == True):
+        applied = AppliedScience.objects.all()
+        health = HealthScience.objects.all()
+        pure = PureScience.objects.all()
+        total_applied = applied.count()
+        total_health = health.count()
+        total_pure = pure.count()
+        total = total_applied + total_health + total_pure
+        context = {
+            'applied': applied,
+            'health': health,
+            'pure': pure,
+            'total': total,
+        }
+        return render(request, 'app_demo_model/show_model.html', context)
+    else:
+        return render(request, 'app_general/errors_page.html')
 
 @login_required
 def delete_all_data(request):
-    applied = AppliedScience.objects.all()
-    health = HealthScience.objects.all()
-    pure = PureScience.objects.all()
-    applied.delete()
-    health.delete()
-    pure.delete()
-    
-    return render(request, 'app_demo_model/show_model.html')
+    user = request.user
+    if (user.is_superuser == True or user.is_staff == True):
+        applied = AppliedScience.objects.all()
+        health = HealthScience.objects.all()
+        pure = PureScience.objects.all()
+        applied.delete()
+        health.delete()
+        pure.delete()
+        return render(request, 'app_demo_model/show_model.html')
+    else:
+        return render(request, 'app_general/errors_page.html')
